requirements.txt
nltk>=3.6.3
numpy>=1.19.5
scikit-learn>=0.24.2

# Web / env
flask>=2.0.0
python-dotenv>=1.0.0

# OpenAI SDK
openai>=1.0.0
